port(runtime): VS.EXE tranche 2 — the battle core, and the seams between its slices
Three surfaces, written in parallel and then integrated: the battle manager
(LAB_80055e3c and its four states, 0x1F1C bytes), the battle scene (LAB_80034eac
and its five phases, 0x2054 bytes), and the roster consumer FUN_8005cbe0, which
turns the six character ids SELECT.EXE exports at 0x801FF102 into a 300-byte
battle setup record. Each agent reported its own seams upward rather than forking
state, which is what made the integration possible at all.

The integration is the substance of this commit, and it found more than the three
reports did.

TEN DECLARATIONS OF ONE ADDRESS, NINE COLLAPSED. A full sweep of every global and
function declared in more than one file turned up two forked globals and ten
forked functions. Four of the ten were not cosmetic: FUN_8003f228, FUN_8003f2b0
and FUN_8003f540 each had a real transliteration in one animation family and EMPTY
STUBS shadowing it in three others, so three opcode families were calling nothing
where a sibling held the code. That shipped two commits ago inside tranche 1, with
a green build. FUN_8005cbe0 was the same shape one layer up: Roster held 118 lines
and main called a local empty stub.

FUN_8003f540 HAD BEEN PORTED TWICE, and neither version had its signature right.
One file typed the switch selector int, the other ushort. The bytes say short:

    0x8003F554  sll   a1,a1,16
    0x8003F558  sra   a1,a1,16      <- sign extension of a halfword
    0x8003F55C  sltiu v0,a1,13

The original truncates its selector to 16 bits before testing the switch range, so
a value like 0x10005 becomes 5 and takes case 5, where the int version fell out of
the switch and did nothing. The two bodies were compared line by line before one
was kept: apart from the type they differed only in parentheses and in rand()
versus Kernel.rand(), which are the same symbol through a using static. Neither was
wrong on the substance. That there were two was the defect.

Return types were settled the same way, from the callers rather than from either
agent's reading: FUN_80042054 (11 callers, 2 read $v0), FUN_800600b0 (4, 1) and
FUN_8005f704 (3, 2) return a value, so the void copies were losing information;
FUN_8005fcec (4 callers, 0) and FUN_8005ee5c (4, 0) do not.

ONE DUPLICATE IS KEPT, DELIBERATELY. FUN_80047550 is declared with two incompatible
parameter lists, and the bytes vindicate both: a1 is a pointer at each call site —
0x801FAA84 built by lui/ori at one, sp+104 at the other. But one C# site passes a
managed VECTOR with no PSX address and the other passes a raw PSX address, and
unifying them means first deciding how this port represents an addressable stack
VECTOR. That is an architecture decision, not a deduction, so both stubs stay, both
now carry the full evidence, and neither pretends the question is closed.

THE MANAGER WAS UNREACHABLE, the third instance this tranche. main created its task
node without registering the callback, so list 9 walked a live node and dispatched
nothing. Same shape as the fighter body last commit and the resolver before it, and
all three sat where two slices met.

So this one gets a guard instead of a lesson. check_task_registration.py asserts
the invariant across all three overlays: every CreateTask entry either has a
registered callback or has no ported body at all. The first version of it was
vacuous — commenting out a registration to check that it failed did not make it
fail, because it was matching the name inside the commented line. A checker that
cannot fail proves nothing, so it now strips comments first, and the negative
control does fail. Twenty task creations across VS, TITLE and SELECT: invariant
holds, with the seven genuinely unported entries reported as BLOCKED rather than
passed over.

Also closed: BattleState now names the context head that three files were spelling
in raw hex, and ctx+0x14 / ctx+0x16 turn out to be one acting-slot cursor per team,
which lifts a PARTIAL FighterTask had left open and was right to leave open — its
own function could not see it.

Build clean. Ten benches green, SELECT.EXE unchanged at 49396 lit pixels, the
51-slot opcode table still matches the binary, cross-file duplicate globals now
zero and duplicate functions one.

+++ custom-tools/DbzLegendsAnalyser/DbzLegendsRemaster/VS_EXE/AnimCmdTransform.cs
@@ -169,7 +169,7 @@ internal static class AnimCmdTransform
                     puVar6 = puVar6 + 2;
                     if (psVar3 != 0)
                     {
-                        sVar2 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar3), uVar7,
+                        sVar2 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar3), (short)(uVar7),
                             (uint)(int)(short)uVar1);
                         PsxRam.WriteU16(psVar3, (ushort)sVar2);
                         if (uVar7 == 8)
@@ -262,7 +262,7 @@ internal static class AnimCmdTransform
                     }
 
                     puVar6 = puVar6 + 2;
-                    sVar2 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar3), uVar7,
+                    sVar2 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar3), (short)(uVar7),
                         (uint)(int)(short)uVar1);
                     PsxRam.WriteU16(psVar3, (ushort)sVar2);
                     if (uVar7 == 8)
@@ -362,7 +362,7 @@ internal static class AnimCmdTransform
                     }
 
                     streamPtr = streamPtr + 2;
-                    sVar2 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(unaff_s4), uVar3,
+                    sVar2 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(unaff_s4), (short)(uVar3),
                         (uint)(int)(short)uVar1);
                     PsxRam.WriteU16(unaff_s4, (ushort)sVar2);
                     if (uVar3 == 8)
@@ -501,7 +501,7 @@ internal static class AnimCmdTransform
                                 do
                                 {
                                     sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar10),
-                                        (ushort)(uStack_3e & 0xf), (uint)(int)(short)auStack_38[iVar8]);
+                                        (short)((ushort)(uStack_3e & 0xf)), (uint)(int)(short)auStack_38[iVar8]);
                                     PsxRam.WriteU16(psVar10, (ushort)sVar3);
                                     psVar10 = psVar10 + 8;
                                     iVar8 = iVar8 + 1;
@@ -531,7 +531,7 @@ internal static class AnimCmdTransform
                             do
                             {
                                 sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar10),
-                                    (ushort)(uStack_3e & 0xf), (uint)(int)(short)auStack_38[iVar7]);
+                                    (short)((ushort)(uStack_3e & 0xf)), (uint)(int)(short)auStack_38[iVar7]);
                                 PsxRam.WriteU16(psVar10, (ushort)sVar3);
                                 psVar10 = psVar10 + 8;
                                 iVar7 = iVar7 + 1;
@@ -567,7 +567,7 @@ internal static class AnimCmdTransform
                                     do
                                     {
                                         sVar4 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar10),
-                                            (ushort)(uStack_3e & 0xf), (uint)(int)(short)auStack_38[iVar9]);
+                                            (short)((ushort)(uStack_3e & 0xf)), (uint)(int)(short)auStack_38[iVar9]);
                                         PsxRam.WriteU16(psVar10, (ushort)sVar4);
                                         psVar10 = psVar10 + 8;
                                         iVar9 = iVar9 + 1;
@@ -676,7 +676,7 @@ internal static class AnimCmdTransform
                                 {
                                     puVar11 = puVar11 + 8;
                                     sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                        (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar9]);
+                                        (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar9]);
                                     PsxRam.WriteU16(psVar8, (ushort)sVar3);
                                     psVar8 = psVar8 + 8;
                                     iVar9 = iVar9 + 1;
@@ -708,7 +708,7 @@ internal static class AnimCmdTransform
                             {
                                 puVar11 = puVar11 + 8;
                                 sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                    (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar6]);
+                                    (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar6]);
                                 PsxRam.WriteU16(psVar8, (ushort)sVar3);
                                 psVar8 = psVar8 + 8;
                                 iVar6 = iVar6 + 1;
@@ -746,7 +746,7 @@ internal static class AnimCmdTransform
                                     {
                                         puVar11 = puVar11 + 8;
                                         sVar4 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                            (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar10]);
+                                            (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar10]);
                                         PsxRam.WriteU16(psVar8, (ushort)sVar4);
                                         psVar8 = psVar8 + 8;
                                         iVar10 = iVar10 + 1;
@@ -853,7 +853,7 @@ internal static class AnimCmdTransform
                                 {
                                     puVar11 = puVar11 + 8;
                                     sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                        (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar9]);
+                                        (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar9]);
                                     PsxRam.WriteU16(psVar8, (ushort)sVar3);
                                     psVar8 = psVar8 + 8;
                                     iVar9 = iVar9 + 1;
@@ -885,7 +885,7 @@ internal static class AnimCmdTransform
                             {
                                 puVar11 = puVar11 + 8;
                                 sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                    (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar6]);
+                                    (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar6]);
                                 PsxRam.WriteU16(psVar8, (ushort)sVar3);
                                 psVar8 = psVar8 + 8;
                                 iVar6 = iVar6 + 1;
@@ -923,7 +923,7 @@ internal static class AnimCmdTransform
                                     {
                                         puVar11 = puVar11 + 8;
                                         sVar4 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                            (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar10]);
+                                            (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar10]);
                                         PsxRam.WriteU16(psVar8, (ushort)sVar4);
                                         psVar8 = psVar8 + 8;
                                         iVar10 = iVar10 + 1;
@@ -1033,7 +1033,7 @@ internal static class AnimCmdTransform
                                 {
                                     puVar11 = puVar11 + 8;
                                     sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                        (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar9]);
+                                        (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar9]);
                                     PsxRam.WriteU16(psVar8, (ushort)sVar3);
                                     psVar8 = psVar8 + 8;
                                     iVar9 = iVar9 + 1;
@@ -1065,7 +1065,7 @@ internal static class AnimCmdTransform
                             {
                                 puVar11 = puVar11 + 8;
                                 sVar3 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                    (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar6]);
+                                    (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar6]);
                                 PsxRam.WriteU16(psVar8, (ushort)sVar3);
                                 psVar8 = psVar8 + 8;
                                 iVar6 = iVar6 + 1;
@@ -1103,7 +1103,7 @@ internal static class AnimCmdTransform
                                     {
                                         puVar11 = puVar11 + 8;
                                         sVar4 = (short)FUN_8003f540((uint)(int)(short)PsxRam.ReadU16(psVar8),
-                                            (ushort)(uStack_46 & 0xf), (uint)(int)(short)auStack_40[iVar10]);
+                                            (short)((ushort)(uStack_46 & 0xf)), (uint)(int)(short)auStack_40[iVar10]);
                                         PsxRam.WriteU16(psVar8, (ushort)sVar4);
                                         psVar8 = psVar8 + 8;
                                         iVar10 = iVar10 + 1;
@@ -1240,7 +1240,24 @@ internal static class AnimCmdTransform
     //
     // The `goto LAB_8003f6a4` shape is the original's and is kept literally: case 6 reaches the
     // common tail with a DIFFERENT value in iVar3 than the fall-through path computes.
-    internal static int FUN_8003f540(uint param_1, ushort param_2, uint param_3)
+    // LE SELECTEUR EST UN `short`, ET NI L'UNE NI L'AUTRE DES DEUX TRANSLITERATIONS NE L'AVAIT.
+    // Cette fonction avait ete portee DEUX FOIS, independamment: ici avec `ushort param_2` et dans
+    // AnimCmdAppearance.cs avec `int param_2`. Les octets tranchent contre les deux:
+    //     0x8003F554  sll   a1,a1,16
+    //     0x8003F558  sra   a1,a1,16      <- extension de SIGNE d'un demi-mot
+    //     0x8003F55C  sltiu v0,a1,13
+    // L'original tronque donc son deuxieme argument a 16 bits AVANT de tester la plage du switch.
+    // Ce n'est pas cosmetique: avec `int`, une valeur comme 0x10005 sort du switch et ne fait rien,
+    // alors que l'original la tronque a 5 et prend le cas 5. Avec `ushort`, 0xFFFF vaut 65535 au
+    // lieu de -1 — meme sortie ici, par chance, puisque le sltiu est non signe.
+    //
+    // Les deux corps ont ete compares ligne a ligne avant d'en garder un: hors ce type, ils ne
+    // divergeaient que par des parentheses et par `rand()` contre `Kernel.rand()`, qui sont le meme
+    // symbole via `using static PsxSdkMonogame.Kernel;`. Le calcul du cas 6 differait de place de
+    // cast (`(int)(a*b*0x10000)` contre `(int)(a*b)*0x10000`) et rend les memes 32 bits de poids
+    // faible en contexte unchecked. Aucune des deux versions n'etait fausse sur le fond; c'est le
+    // fait qu'il y en ait deux qui l'etait.
+    internal static int FUN_8003f540(uint param_1, short param_2, uint param_3)
     {
         short sVar1;
         uint uVar2;
